requirements.txt
pixivpy
apscheduler
pillow
pycryptodome